Fix NativeBool (?)
There was an issue where in standalone build, NativeBool did not work (was true when false was expected). Editor Playmode worked fine or at least looked like it worked.

The problem was solved (?) by changing NativeBool Value type to byte.

Amerkoleci's jolt f8c073b7abe5c635093a1d453741f842f055d737 removed JPH_Bool32 for stdbool.h bool.

+++ Jolt/Native/NativeBool.cs
@@ -5,7 +5,7 @@
     /// </summary>
     internal struct NativeBool
     {
-        public uint Value;
+        public byte Value;
 
         public static implicit operator bool(NativeBool value)
         {
@@ -14,7 +14,7 @@
 
         public static implicit operator NativeBool(bool value)
         {
-            return new NativeBool { Value = value ? 1u : 0u };
+            return new NativeBool { Value = value ? (byte)1 : (byte)0 };
         }
     }
 }
